mise à jour

+++ Back-end/Admission/src/main/java/com/example/demo/service/SelecionneService.java
@@ -1,9 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.models.Admin;
 import com.example.demo.models.Selectionne;
 import org.springframework.http.ResponseEntity;
-import org.springframework.kafka.core.KafkaTemplate;
 
 import java.util.List;
 
@@ -11,9 +9,11 @@ public interface SelecionneService {
 
     void addSelectionne(Selectionne selectionne);
 
-    ResponseEntity<String> selectStudentByAdmin(Long adminId, Long etudiantId);
+    ResponseEntity<String> selectStudentByAdmin(Long adminId, int etudiantId);
 
     List<Selectionne> getSelectionnes();
 
     ResponseEntity<String> deleteSelectionne(Long id);
 }
+
+
